refactor: advanced details metrics
