+++ apps/main/src/lend/components/MarketInformationComposite.tsx
@@ -1,7 +1,7 @@
 import { ChartAndActivityComp } from '@/lend/components/ChartAndActivityComp'
 import { networks } from '@/lend/networks'
 import {
-  AdvancedDetails,
+  AdvancedDetailsMetrics,
   MarketInfoLayout,
   MarketParametersCard,
 } from '@/llamalend/features/market-advanced-information'
@@ -47,7 +47,7 @@ export const MarketInformationComposite = ({ rateType, previewPrices }: MarketIn
         <Card size="small">
           <CardHeader title={t`Advanced Details`} />
           <CardContent component={Stack}>
-            <AdvancedDetails />
+            <AdvancedDetailsMetrics />
             <MarketInfoLayout network={networks[chainId]} />
           </CardContent>
         </Card>
