+++ apps/main/src/llamalend/features/market-advanced-information/index.ts
@@ -2,5 +2,6 @@ export * from './MarketInfoLayout'
 export * from './MarketContractsSection'
 export * from './MarketLoanParameters'
 export * from './MarketParametersSection'
-export * from './AdvancedDetails'
+export * from './AdvancedDetailsMetrics'
+export * from './MarketOverviewDetails'
 export * from './MarketParametersCard'
